Refactor decoder API to not need a `.begin()` call
Now our Symphonia and FFmpeg decoder APIs directly
implement the `Source` trait, instead of being
separate structs that create a `Source` with a
call to `.begin()`.
