--- src/backend/ffmpeg/mod.rs
@@ -1,7 +1,0 @@
-mod decoder;
-mod init;
-mod source;
-
-pub use decoder::FFmpegDecoder;
-pub use init::ffmpeg_init;
-pub use source::FFmpegSource;
