+++ src/backend/mod.rs
@@ -1,6 +1,3 @@
-#[cfg(all(feature = "enable-filesystem", feature = "enable-ffmpeg"))]
-pub mod ffmpeg;
-
 #[cfg(all(feature = "enable-multithreading", feature = "enable-filesystem"))]
 pub mod batch;
 
@@ -13,7 +10,6 @@ pub mod constants;
 pub mod decoder;
 pub mod display;
 pub mod source;
-pub mod symphonia;
 pub mod units;
 
 mod batch_args;
@@ -26,7 +22,6 @@ mod waveform_named_result;
 mod waveform_result;
 
 pub use batch_args::BatchArgs;
-pub use decoder::Decoder;
 pub use errors::Error;
 pub use sample::Sample;
 pub use signal::Signal;
