--- src/backend/symphonia/mod.rs
@@ -1,5 +1,0 @@
-mod decoder;
-mod source;
-
-pub use decoder::SymphoniaDecoder;
-pub use source::SymphoniaSource;
